# PERF: sort_index with an already monotonic MultiIndex