- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.2.0.rst` file if fixing a bug or adding a new feature.

```
import pandas as pd
import numpy as np

mi = pd.MultiIndex.from_product([np.arange(1000)] * 2)
df = pd.DataFrame(np.random.randn(len(mi)), index=mi)

%timeit df.sort_index()

# 100 ms ± 2.22 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  <- main
# 4.17 ms ± 62.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  <- PR
```